refactor: move tests to sensible subfolders
Allows adding further e2e tests for other container orchestrators.
