+++ .github/workflows/test_e2e.yml
@@ -2,7 +2,7 @@ name: Run e2e tests
 on: [pull_request]
 concurrency: ci-${{ github.ref }}
 jobs:
-  test:
+  test-kubernetes:
     runs-on: ubuntu-latest
     strategy:
       # The e2e tests are flaky and often one of the jobs fails. The default setting
@@ -67,4 +67,4 @@ jobs:
         tag=$(skaffold build --tag="e2e-${GITHUB_RUN_ID}-${GITHUB_RUN_NUMBER}" --quiet --output="{{ (index .Builds 0).Tag }}")
         skaffold deploy --images=hetznercloud/hcloud-csi-driver=$tag
 
-        e2etests/run-e2e-tests.sh
+        test/e2e/kubernetes/run-e2e-tests.sh
