+++ test/e2e/kubernetes/.gitignore
