+++ test/integration/cryptsetup_test.go
@@ -1,4 +1,4 @@
-package integrationtests
+package integration
 
 import (
 	"os"
